fix(1487): cleanup terraform, refs: #1487
