+++ infrastructure/server/fam_admin_management_api.tf
@@ -132,7 +132,7 @@ resource "aws_lambda_function" "fam_admin_management_api_function" {
       FC_API_TOKEN = "${var.forest_client_api_api_key}"
       FC_API_BASE_URL = "${var.forest_client_api_base_url}"
 
-      IDIM_PROXY_BASE_URL = "${var.idim_proxy_api_base_url}"
+      IDIM_PROXY_BASE_URL_PROD = "${var.idim_proxy_api_base_url_prod}"
       IDIM_PROXY_API_KEY = "${var.idim_proxy_api_api_key}"
     }
