+++ infrastructure/server/variables_provided.tf
@@ -180,20 +180,10 @@ variable "forest_client_api_base_url" {
   sensitive = true
 }
 
-variable "idim_proxy_api_base_url" {
-  type = string
-  sensitive = true
-}
-
 variable "idim_proxy_api_base_url_prod" {
   type = string
   sensitive = true
   default = "no need in dev and test env"
-}
-
-variable "idim_proxy_api_base_url_test" {
-  type = string
-  sensitive = true
 }
 
 variable "idim_proxy_api_api_key" {
