+++ server/backend/local-dev.env
@@ -10,7 +10,6 @@ COGNITO_USER_POOL_DOMAIN=dev-fam-user-pool-domain
 ENABLE_BCSC_JWKS_ENDPOINT=False
 BCSC_KEY_ID=thisisasecret
 FC_API_TOKEN=thisisasecret
-IDIM_PROXY_BASE_URL_TEST=https://nr-fam-idim-lookup-proxy-test-backend.apps.silver.devops.gov.bc.ca
 IDIM_PROXY_API_KEY=thisisasecret
 GC_NOTIFY_EMAIL_API_KEY=thisisasecret
 TEST_IDIR_USER_GUID=thisisasecret
