+++ terraform/prod/terragrunt.hcl
@@ -39,8 +39,6 @@ generate "prod_tfvars" {
   fam_console_idp_name_bceid = "PROD-BCEIDBUSINESS"
   forest_client_api_base_url = "https://nr-forest-client-api-prod.api.gov.bc.ca/"
   use_override_proxy_endpoints = false
-  idim_proxy_api_base_url = "https://nr-fam-idim-lookup-proxy-prod-backend.apps.silver.devops.gov.bc.ca"
   idim_proxy_api_base_url_prod = "https://nr-fam-idim-lookup-proxy-prod-backend.apps.silver.devops.gov.bc.ca"
-  idim_proxy_api_base_url_test = "https://nr-fam-idim-lookup-proxy-test-backend.apps.silver.devops.gov.bc.ca"
 EOF
 }
